pyproject.toml
[tool.poetry]
name = "pynite-01"
version = "0.1.0"
description = "Pynite Example from Flocode Newsletter Article #034."
authors = ["joreilly86 <joreilly86@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pynitefea = "^0.0.94"
numpy = "^1.26.4"
prettytable = "^3.10.0"
pyvista = {extras = ["all", "trame"], version = "^0.43.9"}
ipywidgets = "^8.1.3"
trame-jupyter-extension = "^2.1.1"
jupyterlab = "^4.2.2"
sympy = "^1.12.1"
scipy = "^1.13.1"
trame = "^3.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
